Move `sdk/manifest` validation-related logic to `sdk/validation/manifest`
