+++ pkg/sdk/dbpopulator/loader.go
@@ -33,7 +33,7 @@ func Group(paths []string) (map[string][]string, error) {
 		if err != nil {
 			return manifests, errors.Wrapf(err, "while reading file from path %s", path)
 		}
-		metadata, err := manifest.UnmarshalManifestMetadata(content)
+		metadata, err := manifest.UnmarshalMetadata(content)
 		if err != nil {
 			return manifests, errors.Wrapf(err, "while unmarshaling manifest content from path %s", path)
 		}
