+++ pkg/sdk/validation/manifest/json_remote_helper.go
