+++ pkg/sdk/validation/manifest/json_remote_implementation.go
