+++ pkg/sdk/validation/manifest/json_remote_interface.go
