+++ pkg/sdk/validation/manifest/json_remote_type.go
