+++ pkg/sdk/validation/manifest/json_type.go
