+++ pkg/sdk/validation/manifest/testdata/invalid-interface.yaml
