+++ pkg/sdk/validation/manifest/testdata/invalid-manifest.yaml
