+++ pkg/sdk/validation/manifest/testdata/invalid-type.yaml
