+++ pkg/sdk/validation/manifest/testdata/invalid-type_json-schema.yaml
