+++ pkg/sdk/validation/manifest/testdata/valid-implementation.yaml
