+++ pkg/sdk/validation/manifest/testdata/valid-interface.yaml
