+++ pkg/sdk/validation/manifest/testdata/valid-type.yaml
