+++ pkg/sdk/validation/manifest/types.go
